feat: 新增头部top navigator bar
