+++ .prettierrc.js
@@ -1,13 +1,5 @@
-/*
- * @Author: mingwei
- * @Date: 2022-04-05 08:39:22
- * @LastEditors: mingwei
- * @LastEditTime: 2022-04-05 08:39:37
- * @FilePath: /react-native-dev-sdk/.prettierrc.js
- * @Description:
- */
 module.exports = {
-  printWidth: 100,
+  printWidth: 200,
   tabWidth: 2,
   // semi: true,
   bracketSpacing: true,
@@ -15,5 +7,5 @@ module.exports = {
   // jsxSingleQuote: true,
   singleQuote: true,
   // trailingComma: 'all',
-  arrowParens: "avoid",
+  arrowParens: 'avoid',
 };
