+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-dev-sdk",
-  "version": "1.1.7",
+  "version": "1.1.8",
   "description": "react-native-dev-sdk",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
