+++ src/components/TopNavigation/type.d.ts
@@ -1,10 +1,11 @@
-import React from 'react';
-declare type StatusBarShapeType = {
+declare namespace ITopNavigator {
+  type StatusBarShapeType = {
     barStyle?: any | 'light-content' | 'default';
     hiddle?: boolean;
     backgroundColor?: string;
-};
-interface topNavigationBarType {
+  };
+
+  interface topNavigationBarProps {
     style?: any;
     title?: string;
     titleView?: React.ReactElement;
@@ -17,6 +18,5 @@ interface topNavigationBarType {
     fontSize?: number;
     animated?: boolean;
     hide?: boolean;
+  }
 }
-declare const TopNavigationBar: React.FC<topNavigationBarType>;
-export default TopNavigationBar;
